Removed autopatching of urlconf
The reason is it only worked in Django versions supporting AppConfig (1.7 and newer)
and is rather hard to test if we decide to create some workaround for older Django versions.
Asking the user to include urls into his urlconf manually seems to be the easiest option.
When doing so the user gets the ability to include that urls to whatever location he wants,
this was also impossible with auto-patching. Well, not quite impossible but requiring some amount of work.